% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SST.R
\name{SST_standalone}
\alias{SST_standalone}
\title{Launch the SST test standalone}
\usage{
SST_standalone(
  item_bank = paste0("custom-assets/stimuli/berkowitz_musicxml/",
    list.files(system.file("www/stimuli/berkowitz_musicxml", package = "SST"), pattern =
    "\\\\.musicxml$")),
  app_name,
  musicassessr_aws = FALSE,
  get_p_id = FALSE,
  data_collection_method = "audio",
  no_repetitions = 5L,
  max_goes_forced = FALSE,
  no_items = 10L
)
}
\arguments{
\item{no_items}{}
}
\description{
Launch the SST test standalone
}
